Lesson 1.2: React DOM And Components
