+++ src/AddTodoForm.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const AddTodoForm = () => {
+  return (
+    <>
+      <form>
+        <label htmlFor="todoTitle">Title: </label>
+        <input id="todoTitle" type="text" name="title" />
+        <button type="submit">Add</button>
+      </form>
+    </>
+  );
+};
+
+export default AddTodoForm;
